requirements.txt
# Django and REST framework
django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0

# Database
psycopg2-binary==2.9.6

# Web and API
requests==2.31.0
websockets==11.0.3
flask-socketio==5.3.4

# Data processing and analysis
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.12.2

# Machine Learning
scikit-learn==1.2.2
tensorflow==2.12.0
torch==2.0.1
stable-baselines3==2.0.0

# NLP and Audio Processing
transformers==4.30.2
nltk==3.8.1
speechrecognition==3.10.0
librosa==0.10.0.post2

# Computer Vision
opencv-python==4.7.0.72
dlib==19.24.1
face-recognition==1.3.0

# Time Series Analysis
statsmodels==0.14.0

# Dashboard and Visualization
streamlit==1.24.0

# Deployment
gunicorn==20.1.0
whitenoise==6.5.0
dj-database-url==2.0.0

# Development and Testing
pytest==7.3.1
black==23.3.0
flake8==6.0.0